+++ .github/workflows/publish.yml
@@ -28,11 +28,22 @@ jobs:
               with:
                   node-version: ${{ matrix.node-version }}
 
-            - name: yarn
-              run: yarn install
+            - name: Cache node modules
+              uses: actions/cache@v3
+              env:
+                cache-name: cached-node-modules
+              with:
+                path: node_modules
+                key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('yarn.lock') }}
 
-            #- name: Test
-            #  run: npm test
+            - name: Install dependencies
+              run: yarn install --frozen-lockfile
+
+            - name: Build
+              run: yarn build
+
+            - name: Test
+              run: yarn test
 
             #- name: Package
             #  run: yarn package
