+++ package.json
@@ -18,21 +18,27 @@
         "install": "node-pre-gyp install --fallback-to-build=false || yarn build:rs:release",
         "watch": "tsc -w -p tsconfig.json",
         "package": "node-pre-gyp package",
+        "test": "jest",
         "upload-binary": "yarn build && yarn node-pre-gyp package && yarn node-pre-gyp-github publish",
         "prepack": "shx rm -rf bin"
     },
     "devDependencies": {
+        "@jest/globals": "^29.4.3",
         "@mathquis/node-pre-gyp-github": "1.0.1",
+        "@types/jest": "^29.4.0",
         "@types/mocha": "^5.2.7",
         "@types/node": "12.12.54",
         "chai": "^4.0.1",
         "chai-as-promised": "^7.1.0",
         "chai-bignumber": "^3.0.0",
         "ethereum-types": "^3.4.1",
+        "ethers": "^6.0.4",
+        "jest": "^29.4.3",
         "mocha": "^6.2.0",
         "shx": "^0.3.3",
+        "ts-jest": "^29.0.5",
         "tslint": "5.11.0",
-        "typescript": "4.2.2"
+        "typescript": "4.9.5"
     },
     "repository": {
         "type": "git",
