+++ src/__tests__/BalanceChecker.abi.json
@@ -1,0 +1,31 @@
+[
+	{
+	 "payable": true,
+	 "stateMutability": "payable",
+	 "type": "fallback"
+	},
+	{
+	 "constant": true,
+	 "inputs": [
+	  {
+	   "name": "user",
+	   "type": "address"
+	  },
+	  {
+	   "name": "token",
+	   "type": "address"
+	  }
+	 ],
+	 "name": "tokenBalance",
+	 "outputs": [
+	  {
+	   "name": "",
+	   "type": "uint256"
+	  }
+	 ],
+	 "payable": false,
+	 "stateMutability": "view",
+	 "type": "function"
+	}
+   ]
+
