+++ tsconfig.json
@@ -3,6 +3,7 @@
         "outDir": "lib",
         "rootDir": "src",
         "resolveJsonModule": true,
+        "esModuleInterop": true,
         "module": "commonjs",
         "target": "es6",
         "lib": ["es2017", "esnext.asynciterable", "es2018.promise"],
